[REFACTORING] Split IMAPServerPlainAuthEnabledWithoutRequireSSLTest
